Introduce TaxProcessor
Now that the Product class is returning the correct prices with sales
tax and import duty applied appropriately, we introduce a TaxProcessor
class, with a supporting LineItem class so that a receipt can be
generated. A spec for the first shopping basket demonstrates that the
receipt is generated as expected.